pyproject.toml
[project]
name = "nomad"
version = "0.2.1"
description = "Capture, detect and notify from camera live feed."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pydantic-settings>=2.7.1",
    "python-telegram-bot>=21.10",
    "ultralytics>=8.3.70",
]

[tool.setuptools]
py-modules = []
packages = ["nomad", "nomad.modules"]

[project.scripts]
nomad = "nomad.__main__:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
